C (#44)
* C: 124, just dropping whitespace

* C: 104 bytes

gcc will infer an int return type and 0 return, drop those, though it generates a warning (but not an error).

main takes arguments argc and argv. argc is 1 when the program is invoked without arguments beyond the program name. use this to skip initialization of i. it's wrong if you try to pass arguments to the program, but that's fine.

one more byte changing <= to <.